feat(ui): move Export/Import to the top query toolbar
Relocate the Export and Import buttons out of the results toolbar (next to
Insert/Schema) up into the query action toolbar (Run / Load query row),
beside AI Helper / Visual Query Builder. Handlers now flow App -> DocumentViewer
instead of App -> DataGrid; DataGrid no longer owns those buttons.
