+++ src/components/DataGrid.tsx
@@ -1,7 +1,7 @@
 import React, { useState, useMemo, useEffect, useContext } from 'react';
 import { DocumentViewerContext } from './DocumentViewer';
 import { List } from 'react-window';
-import { Table, Braces, ChevronRight, ChevronDown, ListFilter, Copy, Check, Edit, Trash2, Plus, Download, Upload, Table2, BarChart3 } from 'lucide-react';
+import { Table, Braces, ChevronRight, ChevronDown, ListFilter, Copy, Check, Edit, Trash2, Plus, Table2, BarChart3 } from 'lucide-react';
 import { ChartView } from './ChartView';
 import { EJSON, ObjectId, Long, Decimal128, Int32, Double, Binary, Timestamp } from 'bson';
 
@@ -15,8 +15,6 @@ interface DataGridProps {
   onInsertDocument?: () => void;
   onEditDocument?: (doc: Record<string, any>) => void;
   onDeleteDocument?: (doc: Record<string, any>) => void;
-  onOpenExport?: () => void;
-  onImport?: () => void;
   onAnalyzeSchema?: () => void;
   onUpdateMany?: () => void;
   onDeleteMany?: () => void;
@@ -325,8 +323,6 @@ export const DataGrid: React.FC<DataGridProps> = ({
   onInsertDocument,
   onEditDocument,
   onDeleteDocument,
-  onOpenExport,
-  onImport,
   onAnalyzeSchema,
   onUpdateMany,
   onDeleteMany,
@@ -988,29 +984,6 @@ export const DataGrid: React.FC<DataGridProps> = ({
             >
               <Plus size={12} />
               <span>Insert</span>
-            </button>
-          )}
-          {activeTab === 'results' && onOpenExport && (
-            <button
-              type="button"
-              onClick={onOpenExport}
-              className="flex items-center gap-1.5 px-2 py-1 rounded text-[11px] font-medium border border-[var(--border-color)] text-[var(--text-muted)] hover:text-[var(--text-main)] hover:bg-[var(--bg-item-hover)] cursor-pointer transition-all"
-              title="Open export workspace"
-              data-testid="export-btn"
-            >
-              <Download size={12} />
-              <span>Export</span>
-            </button>
-          )}
-          {activeTab === 'results' && onImport && (
-            <button
-              onClick={onImport}
-              className="flex items-center gap-1.5 px-2 py-1 rounded text-[11px] font-medium border border-[var(--border-color)] text-[var(--text-muted)] hover:text-[var(--text-main)] hover:bg-[var(--bg-item-hover)] cursor-pointer transition-all"
-              title="Import documents from a file"
-              data-testid="import-btn"
-            >
-              <Upload size={12} />
-              <span>Import</span>
             </button>
           )}
           {activeTab === 'results' && onAnalyzeSchema && (
